pyproject.toml
[tool.poetry]
name = "search-test"
version = "0.1.0"
description = ""
authors = ["Nuung <qlgks1@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2.9.10"
openai = "^1.75.0"
fastapi = {extras = ["all"], version = "^0.116.1"}
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
sentence-transformers = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"